move Noam from trainer into scheduler.py
