+++ .flake8
@@ -11,6 +11,7 @@ per-file-ignores =
     valle/utils/__init__.py: F401,
     valle/__init__.py: F401,
     valle/modules/__init__.py: F401,
+    valle/modules/scheduler.py: F401,
     valle/modules/embedding.py: E203,
     valle/bin/tokenizer.py: E501,
     valle/modules/transformer.py: E501,
